新增gateway模块,并将原有的跨域配置放入gateway;service_hospital模块新增HospApiController

+++ service/service_hospital/src/main/java/com/sub/hosp/controller/api/ApiController.java
@@ -107,6 +107,13 @@ public class ApiController {
         String depcode = (String) paramMap.get("depcode");
 
         // todo 签名
+        String sign = (String) paramMap.get("sign");
+        String key = hospitalSetService.getSignKey(hoscode);
+        String signKeyMD5 = MD5.encrypt(key);
+
+        if (!sign.equals(signKeyMD5)) {
+            throw new YyghException(ResultCodeEnum.SIGN_ERROR);
+        }
 
         departmentService.remove(hoscode, depcode);
